refactor(views): refactor static views into a single file for easier reasoning
