--- config/views/accessibility_statement.py
@@ -1,11 +1,0 @@
-from .template_view_with_context import TemplateViewWithContext
-
-
-class AccessibilityStatementView(TemplateViewWithContext):
-    template_name = "pages/accessibility_statement.html"
-    page_title = "Accessibility statement"
-
-    def get_context_data(self, **kwargs):
-        context = super().get_context_data(**kwargs)
-        context["feedback_survey_type"] = "support"
-        return context
