--- config/views/help_and_guidance.py
@@ -1,11 +1,0 @@
-from .template_view_with_context import TemplateViewWithContext
-
-
-class HelpAndGuidanceView(TemplateViewWithContext):
-    template_name = "pages/help_and_guidance.html"
-    page_title = "Help and guidance"
-
-    def get_context_data(self, **kwargs):
-        context = super().get_context_data(**kwargs)
-        context["feedback_survey_type"] = "help_and_guidance"
-        return context
