--- config/views/how_to_search_find_case_law.py
@@ -1,11 +1,0 @@
-from .template_view_with_context import TemplateViewWithContext
-
-
-class HowToSearchFindCaseLawView(TemplateViewWithContext):
-    template_name = "pages/how_to_search_find_case_law.html"
-    page_title = "How to search Find Case Law"
-
-    def get_context_data(self, **kwargs):
-        context = super().get_context_data(**kwargs)
-        context["feedback_survey_type"] = "how_to_search_find_case_law"
-        return context
