--- config/views/terms_of_use.py
@@ -1,11 +1,0 @@
-from .template_view_with_context import TemplateViewWithContext
-
-
-class TermsOfUseView(TemplateViewWithContext):
-    template_name = "pages/terms_of_use.html"
-    page_title = "Terms of Use"
-
-    def get_context_data(self, **kwargs):
-        context = super().get_context_data(**kwargs)
-        context["feedback_survey_type"] = "support"
-        return context
