--- config/views/understanding_judgements_and_decisions.py
@@ -1,11 +1,0 @@
-from .template_view_with_context import TemplateViewWithContext
-
-
-class UnderstandingJudgmentsAndDecisionsView(TemplateViewWithContext):
-    template_name = "pages/understanding_judgments_and_decisions.html"
-    page_title = "Understanding judgments and decisions"
-
-    def get_context_data(self, **kwargs):
-        context = super().get_context_data(**kwargs)
-        context["feedback_survey_type"] = "understanding_judgments_and_decisions"
-        return context
